Solution with DataSource
