+++ Lesson17HW/Lesson17HW/Source/DataSource/TableAdapterDataSourceOutput.swift
@@ -1,0 +1,14 @@
+//
+//  TableAdapterDataSourceOutput.swift
+//  Lesson17HW
+//
+//  Created by Алина Власенко on 22.04.2024.
+//
+
+import Foundation
+
+protocol TableAdapterDataSourceOutput {
+    
+    func loadData()
+    func setupSectionAdapters() -> TableViewAdapterInput
+}
